requirements.txt
requests==2.24.0
python-telegram-bot==13.3